auth_service/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
werkzeug==3.0.1

